@pm47 2021-05-17
Added more intermediate steps which show that splitting the `local_channels` in columns and only updating the `commitments` part do not change the results.